progress bar moved to common module
